changed app structure and configuration files
