+++ .gitignore
@@ -1,0 +1,18 @@
+__pycache__
+venv/*
+.env
+.sqlite
+
+# *.pyc
+# *.pem
+# *.pub
+# *.tar.gx
+# *.zip
+# *.sql
+# *.db
+# secrets.txt
+# ./tmp
+# ./build/*
+# ./idea/*
+# .idea
+
