+++ catalogue/account/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+account = Blueprint('account', __name__)
+
+from . import views
