+++ catalogue/asset/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+asset = Blueprint('asset', __name__)
+
+from . import views
