+++ catalogue/catalogue/templates/asset.html
@@ -9,7 +9,7 @@
     <div class="graph-form agile_info_shadow">
          <h3 class="w3_inner_tittle two">New Asset </h3>
             <div class="form-body">
-                <form method="POST" action="{{ url_for('assets.asset') }}">
+                <form method="POST" action="{{ url_for('asset.asset') }}">
 
                     {{ form.hidden_tag() }}
                             <!--Serial Number-->
