+++ catalogue/catalogue/templates/layout.html
@@ -32,7 +32,7 @@ Smartphone Compatible web template, free webdesigns for Nokia, Samsung, LG, Sony
 					<ul id="gn-menu" class="gn-menu-main">
 
 						<!-- //nav_agile_w3l -->
-		                <li class="second logo admin"><h1><a href="main-page.html"><i class="fa fa-graduation-cap" aria-hidden="true"></i>ASSET INVENTORY </a></h1></li>				
+		                <li class="second logo admin"><h1><a href="main_page.html"><i class="fa fa-graduation-cap" aria-hidden="true"></i>ASSET INVENTORY </a></h1></li>				
 
 					</ul>
 					<!-- //nav -->
