+++ catalogue/catalogue/templates/login.html
@@ -75,7 +75,7 @@
 													</div>
 											
 												
-													 <h4>Don't have an account?<a href="{{ url_for('users.register') }}"> Register</a><h6>
+													 <h4>Don't have an account?<a href="{{ url_for('account.register') }}"> Register</a><h6>
 
 													 
 												</div>
