+++ requirements.txt
@@ -1,0 +1,23 @@
+Babel==2.8.0
+bcrypt==3.1.7
+cffi==1.14.0
+click==7.1.1
+dnspython==1.16.0
+email-validator==1.0.5
+Flask==1.1.2
+Flask-Babel==1.0.0
+Flask-Bcrypt==0.7.1
+Flask-Login==0.5.0
+Flask-SQLAlchemy==2.4.1
+Flask-Table==0.5.0
+Flask-WTF==0.14.3
+idna==2.9
+itsdangerous==1.1.0
+Jinja2==2.11.2
+MarkupSafe==1.1.1
+pycparser==2.20
+pytz==2019.3
+six==1.14.0
+SQLAlchemy==1.3.16
+Werkzeug==1.0.1
+WTForms==2.3.1
